backend/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
datasets==2.15.0
dill==0.3.7
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.104.1
filelock==3.13.1
frozenlist==1.4.0
fsspec==2023.10.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
huggingface-hub==0.19.4
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
loguru==0.7.2
MarkupSafe==2.1.3
mmh3==3.1.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.2.1
nltk==3.8.1
numpy==1.25.2
orjson==3.9.10
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
pinecone-client==2.2.4
pinecone-text==0.7.0
protobuf==3.20.3
pyarrow==14.0.1
pyarrow-hotfix==0.6
pydantic==2.5.2
pydantic-core==2.14.5
pydantic-extra-types==2.1.0
pydantic-settings==2.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
threadpoolctl==3.2.0
tokenizers==0.15.0
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.1
transformers==4.35.2
typing-extensions==4.8.0
tzdata==2023.3
ujson==5.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
watchfiles==0.21.0
websockets==12.0
wget==3.2
win32-setctime==1.1.0
xxhash==3.4.1
yarl==1.9.3
